API1: renamed generic name 'util' into specific 'fixp'
