+++ api1/include/redpitaya/asg_per.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-#include "redpitaya/util.h"
+#include "redpitaya/fixp.h"
 
 #define CWM 14
 #define CWF 16
