+++ api1/include/redpitaya/osc_trg.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-#include "redpitaya/util.h"
+#include "redpitaya/fixp.h"
 
 enum edge {pos, neg};
